refactor is_port_open
